requirements/prod.txt
alembic==1.0.11
click==7.0
cropresize2==0.2.0
flask-mako==0.4
flask-migrate==2.5.2
flask-sqlalchemy==2.4.0
flask==1.1.1
itsdangerous==1.1.0
jinja2==2.10.1
mako==1.1.0
markupsafe==1.1.1
pillow==6.1.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
short-url==1.2.2
six==1.12.0
sqlalchemy==1.3.7
werkzeug==0.15.5
